switches to usize for hash table data for the moment

+++ src/types.rs
@@ -3,8 +3,8 @@ pub type Exponent = u32;
 
 // Hash table stuff
 pub type HashValue = usize;
-pub type HashTableLength = u32;
-pub type DivisorMask = u32;
+pub type HashTableLength = usize;
+pub type DivisorMask = usize;
 
 // Field stuff
 pub type Characteristic = i32;
